TUC-138 Added tests for SampleExtractor
It validates the units handling and dates handlings

This commit also contains minor cosmetic modifs to other files

	modified:   make/qtcreator/tucutestcore/tucutestcore.pro
	modified:   software_documentation/source/drug2_description.rst
	modified:   src/tucucore/montecarlopercentilecalculator.cpp
	modified:   src/tucucore/sampleextractor.cpp
	modified:   src/tucucore/sampleextractor.h
		* Documented the function and made it non static
	modified:   test/tucucore/test_covariateextractor.h
	new file:   test/tucucore/test_sampleextractor.h
		* The new tests
	modified:   test/tucucore/tests.cpp

+++ test/tucucore/test_covariateextractor.h
@@ -1252,7 +1252,7 @@ struct TestCovariateExtractor : public fructose::test_base<TestCovariateExtracto
 
 //        printCovariateSeries(series);
 
-        fructose_assert_eq(series.size() , 7);
+        fructose_assert_eq(series.size() , static_cast<size_t>(7));
         fructose_assert(rc == CovariateExtractor::Result::Ok);
         // First measure propagated back to the start of the interval.
         fructose_assert(covariateEventIsPresent("Gist",
